Extract outputs list into component for reuse
