+++ .flowconfig
@@ -3,7 +3,6 @@
 .*/dist/.*
 
 [include]
-../cashlib/lib/
 
 [libs]
