+++ package.json
@@ -47,7 +47,7 @@
     "colors": "^1.3.2",
     "node-localstorage": "^1.3.1",
     "stringfu": "0.0.2",
-    "tooey": "0.0.1"
+    "tooey": "0.0.2"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
